+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Fixed
 * Viewer Grid: Screenreader spreekt titel van `filter-panel` niet juist uit ([#3583](https://github.com/dso-toolkit/dso-toolkit/issues/3583))
+* Advanced Select: Type `AdvancedSelectOption<...>` is not assignable to type `AdvancedSelectOption<never>` ([#3632](https://github.com/dso-toolkit/dso-toolkit/issues/3632))
 
 ## 🥞 Release 91.1.0 - 2026-04-07